Fix build issue (+ oops)
This is part of a bigger issue. There may be more issues related to
PrismLauncher having identical name of repo & organization.

Changing identifier/namespace/domain whatever you call it, ain't gonna be
so easy, until we make helper variable and cautiously verify all
occurences (Waiting for CI/CD to fail is not the best.)